:truck: rename layout to add other layout
